feat: pretty print transfer
