+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/vegaprotocol/vega-snap.git"
   },
   "source": {
-    "shasum": "IyPQof39Tm5ook2nbdgrZEch5pnEFXByRVlzOoZfFIQ=",
+    "shasum": "YyUt0pynBuGsjekCHeY48oAXiwdaoowLIidfgYTRE+k=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
